test: keep math collection free of product-boundary imports
Move Catalog and dispatch boots for Laurent multiply, autocorrelation,
and unpublished record-minima into tests/catalog.
