+++ tests/math/polynomials/test_laurent_multiply.py
@@ -3,7 +3,6 @@
 import pytest
 
 from jacobian._exact import CanonicalRational
-from jacobian.catalog.catalog import Catalog
 from jacobian.catalog.models import (
     OperationDomainValidationError,
     OperationResourceAdmissionError,
@@ -80,30 +79,6 @@ def test_exponent_growth_is_rejected_before_convolution() -> None:
     )
     with pytest.raises(OperationResourceAdmissionError) as error:
         rational_laurent_multiply(left, right)
-    assert error.value.errors()[0]["type"] == "polynomial.laurent.exponent_growth"
-
-
-def test_catalog_admits_exponent_growth_before_convolution_bound() -> None:
-    operation = Catalog.open().operation("polynomial.laurent.rational.multiply.compute")
-    assert operation is not None
-
-    left_count = 65
-    right_count = 64
-    left = RationalLaurentPolynomial(
-        variables=("x",),
-        terms=tuple(
-            term(1, MAX_POLYNOMIAL_EXPONENT - index) for index in range(left_count)
-        ),
-    )
-    right = RationalLaurentPolynomial(
-        variables=("x",),
-        terms=tuple(term(1, index) for index in range(right_count - 1, -1, -1)),
-    )
-    request = operation.request_type(left=left, right=right)
-
-    with pytest.raises(OperationResourceAdmissionError) as error:
-        operation.run(request)
-
     assert error.value.errors()[0]["type"] == "polynomial.laurent.exponent_growth"
